remove resize logic and scrollbar
