+++ ui/src/lib/features/catalog/Catalog.svelte
@@ -26,38 +26,33 @@
   });
 </script>
 
-<div class="flex h-full w-full overflow-hidden">
-  <Sidebar routes={['/apps']} />
-  <div class="container flex h-full flex-shrink flex-col overflow-y-auto py-9">
-    <div class="mx-auto w-full max-w-[1200px] gap-y-9 px-4 sm:px-6 lg:px-8">
-      <div class="catalog-header mb-8">
-        <div class="text-2xl font-semibold leading-9">Applications Deployable on UDS</div>
-        <div class="text-lg font-normal">
-          Secure mission applications that can be deployed anywhere with Unicorn Delivery Service.
-        </div>
-      </div>
-      {#if isLoading}
-        <div class="flex h-full items-center justify-center">
-          <img src="{base}/images/dougandserver.svg" alt="UDS Logo" class="h-20 w-20" />
-          <p class="ml-4 text-lg">Loading applications...</p>
-        </div>
-      {:else if error}
-        <Alert color="danger" title="Error" message="Failed to fetch applications"></Alert>
-      {:else if applications.length === 0}
-        <Alert
-          color="info"
-          title="No applications found"
-          message={`The search query "${applicationStore.getSearchQuery()}" did not match any known applications.`}
-        />
-      {:else}
-        <div class="catalog-apps flex flex-wrap justify-center gap-2 md:justify-start md:gap-6">
-          {#each applications as app}
-            <div class="mb-4 w-full sm:w-auto">
-              <AppCard {app} />
-            </div>
-          {/each}
-        </div>
-      {/if}
+<div class="flex h-full flex-col overflow-y-auto py-9 px-24">
+  <div class="mb-8">
+    <div class="text-2xl font-semibold leading-9">Applications Deployable on UDS</div>
+    <div class="text-lg font-normal">
+      Secure mission applications that can be deployed anywhere with Unicorn Delivery Service.
     </div>
   </div>
+  {#if isLoading}
+    <div class="flex h-full items-center justify-center">
+      <img src="{base}/images/dougandserver.svg" alt="UDS Logo" class="h-20 w-20" />
+      <p class="ml-4 text-lg">Loading applications...</p>
+    </div>
+  {:else if error}
+    <Alert color="danger" title="Error" message="Failed to fetch applications"></Alert>
+  {:else if applications.length === 0}
+    <Alert
+      color="info"
+      title="No applications found"
+      message={`The search query "${applicationStore.getSearchQuery()}" did not match any known applications.`}
+    />
+  {:else}
+    <div class="catalog-apps flex flex-wrap justify-center gap-2 md:justify-start md:gap-6">
+      {#each applications as app}
+        <div class="mb-4 w-full sm:w-auto">
+          <AppCard {app} />
+        </div>
+      {/each}
+    </div>
+  {/if}
 </div>
